Fix audit issues: split oversized test files, add annotations, scrollbar, doc links
- Split chat_view/tests.rs (1019 lines) into tests.rs + scroll_tests.rs
- Split calendar/tests.rs (1004 lines) into tests.rs + view_tests.rs
- Add annotation registration to flame_graph and dependency_graph render
- Add scrollbar rendering to log_correlation stream panels
- Remove redundant (super::Disableable) and (super::Focusable) doc link
  paths in 9 component files where the traits are already imported
