+++ src/component/chat_view/tests.rs
@@ -363,82 +363,6 @@ fn test_focus_history() {
 }
 
 // =============================================================================
-// Scrolling
-// =============================================================================
-
-#[test]
-fn test_scroll_down() {
-    let mut state = state_with_messages();
-    state.set_auto_scroll(false);
-    state.scroll.set_offset(0);
-    ChatView::update(&mut state, ChatViewMessage::ScrollDown);
-    assert_eq!(state.scroll_offset(), 1);
-}
-
-#[test]
-fn test_scroll_up() {
-    let mut state = state_with_messages();
-    state.set_auto_scroll(false);
-    state.scroll.set_content_length(state.messages().len());
-    state.scroll.set_offset(2);
-    ChatView::update(&mut state, ChatViewMessage::ScrollUp);
-    assert_eq!(state.scroll_offset(), 1);
-}
-
-#[test]
-fn test_scroll_up_at_top() {
-    let mut state = state_with_messages();
-    state.scroll.set_offset(0);
-    ChatView::update(&mut state, ChatViewMessage::ScrollUp);
-    assert_eq!(state.scroll_offset(), 0);
-}
-
-#[test]
-fn test_scroll_to_top() {
-    let mut state = state_with_messages();
-    state.scroll.set_content_length(state.messages().len());
-    state.scroll.set_offset(2);
-    ChatView::update(&mut state, ChatViewMessage::ScrollToTop);
-    assert_eq!(state.scroll_offset(), 0);
-    assert!(!state.auto_scroll());
-}
-
-#[test]
-fn test_scroll_to_bottom() {
-    let mut state = state_with_messages();
-    state.set_auto_scroll(false);
-    state.scroll.set_offset(0);
-    ChatView::update(&mut state, ChatViewMessage::ScrollToBottom);
-    assert!(state.scroll.at_end());
-    assert!(state.auto_scroll());
-}
-
-#[test]
-fn test_auto_scroll_on_new_message() {
-    let mut state = focused_state();
-    state.push_user("a");
-    state.push_user("b");
-    // auto_scroll should keep us at the bottom
-    assert!(state.auto_scroll());
-}
-
-#[test]
-fn test_scroll_up_disables_auto_scroll() {
-    let mut state = state_with_messages();
-    ChatView::update(&mut state, ChatViewMessage::ScrollUp);
-    assert!(!state.auto_scroll());
-}
-
-#[test]
-fn test_set_auto_scroll() {
-    let mut state = ChatViewState::new();
-    state.set_auto_scroll(false);
-    assert!(!state.auto_scroll());
-    state.set_auto_scroll(true);
-    assert!(state.auto_scroll());
-}
-
-// =============================================================================
 // Disabled state
 // =============================================================================
 
@@ -820,13 +744,6 @@ fn test_submit_clears_input() {
     ChatView::update(&mut state, ChatViewMessage::Input('x'));
     ChatView::update(&mut state, ChatViewMessage::Submit);
     assert!(state.input_value().is_empty());
-}
-
-#[test]
-fn test_scroll_empty_history() {
-    let mut state = focused_state();
-    ChatView::update(&mut state, ChatViewMessage::ScrollDown);
-    assert_eq!(state.scroll_offset(), 0);
 }
 
 #[test]
