+++ src/component/data_grid/mod.rs
@@ -5,7 +5,7 @@
 //! Enter again to confirm the edit. State is stored in [`DataGridState<T>`],
 //! updated via [`DataGridMessage`], and produces [`DataGridOutput`].
 //!
-//! Implements [`Focusable`] and [`Disableable`](super::Disableable).
+//! Implements [`Focusable`] and [`Disableable`].
 //!
 //! # Example
 //!
