+++ src/component/event_stream/mod.rs
@@ -8,7 +8,7 @@
 //! State is stored in [`EventStreamState`], updated via [`EventStreamMessage`],
 //! and produces [`EventStreamOutput`].
 //!
-//! Implements [`Focusable`] and [`Disableable`](super::Disableable).
+//! Implements [`Focusable`] and [`Disableable`].
 //!
 //! # Example
 //!
